requirements.txt
streamlit
pydantic
pillow
torch
transformers
opencv-python
pandas
